telemetry: split attitude into Euler + Quaternion
Before we processed whatever attitude was being sent which meant that
both MAVLink messages ATTITUDE and ATTITUDE_QUATERNION would trigger
both callbacks. This was often ok but can be very confusing when trying
to set the rate of the messages.

This commit decouples this magic and provides setters for both the
rates.